+++ R/step_spline_convex.R
@@ -1,0 +1,4 @@
+#' @export
+orbital.step_spline_convex <- function(x, all_vars, ...) {
+  spline_helper(x, all_vars, splines2::cSpline)
+}